// src/interfaces.ts
export interface ValidationRule {
  type: 'required' | 'isNumeric' | 'isValidDate' | 'isDni' | 'isRuc' | 'maxLength' | 'minLength';
  value?: number;
  message: string;
}